refactor: create util method to validate URIs
