+++ clients/spring-boot-starter-camunda-sdk/src/test/java/io/camunda/spring/client/properties/CamundaClientPropertiesPostProcessorTest.java
@@ -16,7 +16,6 @@
 package io.camunda.spring.client.properties;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 import java.net.URI;
 import java.time.Duration;
@@ -56,30 +55,6 @@ public class CamundaClientPropertiesPostProcessorTest {
     void shouldMapLists() {
       assertThat(camundaClientProperties.getWorker().getDefaults().getTenantIds())
           .contains("<default>", "another one");
-    }
-  }
-
-  @Nested
-  class InvalidPropertiesTest {
-
-    @Test
-    void shouldThrowExceptionOnInvalidGrpcURI() {
-      assertThatThrownBy(
-              () -> {
-                new CamundaClientProperties().setGrpcAddress(new URI("invalid:26500"));
-              })
-          .isInstanceOf(IllegalArgumentException.class)
-          .hasMessageContaining("The gRPC address must be an absolute URI");
-    }
-
-    @Test
-    void shouldThrowExceptionOnInvalidRestURI() {
-      assertThatThrownBy(
-              () -> {
-                new CamundaClientProperties().setRestAddress(new URI("invalid:8080"));
-              })
-          .isInstanceOf(IllegalArgumentException.class)
-          .hasMessageContaining("The REST address must be an absolute URI");
     }
   }
